pyproject.toml
[tool.poetry]
name = "test-multi-cloud-cicd"
version = "0.1.0"
description = ""
authors = ["codeanish <anish@codeanish.com>"]
readme = "README.md"
packages = [{include = "test_multi_cloud_cicd"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
